Version 4.1
4.0 specs
Game Running with all required specs.
Jar working but cannot upload for some reason which seems to have
occurred with the first project so I will be including a link below to
a google drive folder with both jars.

https://drive.google.com/drive/folders/0Bx5Fjdp3VEcWQUxTSUUzcnh5aUU?usp=sharing

4.1 specs
Changed way images are loaded so that images will only be loaded once significantly increasing the speed of starting a new game. The previous
version worked perfectly fine I just got bored and wanted to up the speed.
The new class for loading the images is not in the design diagram. I had to
upload a new jar file to google drive and upload the project again because
I accidentally deleted the background issue while removing unused images.